+++ changelog.md
@@ -5,6 +5,7 @@ Summary of implemented features classified by release (github tag)
 ## [0.3.0] - 2020-04-30
 
 - fix issue #67 : implement post update + tests
+- fix issue #51 : implement post delete + tests
 
 ## [0.2.5] - 2020-04-28